Use more specific dfw_sst.bin for sklnau8825adi
This is for caroline/chell

This allows for coexistence with differing dfw_sst.bin files needed
by other devices (e.g. Sentry)

The "9d70..." name is what is loaded by the kernel module, and is
a symlink to a renamed dfw_sst.bin.skl... name, which is left
around to help humans looking for the more common dfw_sst.bin.